redirect controller, service and tests

+++ src/main/java/anna/freimuth/urlshortener/repo/UrlRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UrlRepo extends CrudRepository<Url,Integer> {
+public interface UrlRepo extends CrudRepository<Url,Long> {
 }
